reorganize bounty page folder structure
