+++ src/containers/Bounty/components/index.js
@@ -1,14 +1,14 @@
-import ExtendDeadlineErrorModal from './ExtendDeadlineErrorModal';
+import ExtendDeadlineErrorModal from './modals/ExtendDeadlineErrorModal';
 import ModalManager from './ModalManager';
-import ActivateDraftFormModal from './ActivateDraftFormModal';
-import ExtendDeadlineFormModal from './ExtendDeadlineFormModal';
-import ActivateDeadFormModal from './ActivateDeadFormModal';
-import IncreasePayoutFormModal from './IncreasePayoutFormModal';
-import SubmissionItem from './SubmissionItem';
-import FulfillBountyFormModal from './FulfillBountyFormModal';
-import KillBountyFormModal from './KillBountyFormModal';
+import ActivateDraftFormModal from './modals/ActivateDraftFormModal';
+import ExtendDeadlineFormModal from './modals/ExtendDeadlineFormModal';
+import ActivateDeadFormModal from './modals/ActivateDeadFormModal';
+import IncreasePayoutFormModal from './modals/IncreasePayoutFormModal';
+import SubmissionItem from './listItems/SubmissionItem';
+import FulfillBountyFormModal from './modals/FulfillBountyFormModal';
+import KillBountyFormModal from './modals/KillBountyFormModal';
 import NewCommentForm from './NewCommentForm';
-import CommentItem from './CommentItem';
+import CommentItem from './listItems/CommentItem';
 
 export {
   ExtendDeadlineErrorModal,
