+++ src/containers/Bounty/components/listItems/CommentItem.module.scss
