+++ src/containers/Bounty/components/listItems/SubmissionItem.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styles from './SubmissionItem.module.scss';
+import FontAwesomeIcon from '@fortawesome/react-fontawesome';
 import { Button, Card, ListGroup, Text } from 'components';
 import { FulfillmentStagePill, LinkedAvatar } from 'explorer-components';
 import moment from 'moment';
@@ -76,6 +77,10 @@ const SubmissionItem = props => {
               <Text color="defaultGrey" className={styles.label}>
                 Associated files
               </Text>
+              <FontAwesomeIcon
+                icon={['fal', 'file-archive']}
+                className={styles.fileIcon}
+              />
               <Text
                 link
                 src={`https://ipfs.infura.io/ipfs/${dataHash}/${dataFileName}`}
