+++ src/containers/Bounty/components/listItems/SubmissionItem.module.scss
@@ -23,3 +23,7 @@
   flex-direction: column;
   align-items: flex-end;
 }
+
+.fileIcon {
+  margin-right: $base-spacing;
+}
