+++ src/containers/Bounty/components/modals/ActivateDeadFormModal.js
