+++ src/containers/Bounty/components/modals/KillBountyFormModal.js
