+++ src/containers/Bounty/components/modals/Modals.module.scss
